+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/timeline/SoundStreamFrameRange.java
@@ -80,7 +80,7 @@ public class SoundStreamFrameRange implements TreeItem, SoundTag {
 
     @Override
     public boolean importSupported() {
-        return false; //??
+        return true;
     }
 
     @Override