+++ CHANGELOG.md
@@ -27,6 +27,7 @@ All notable changes to this project will be documented in this file.
 - AS3 - navigation to definition in other SWF file and also player/airglobal
 - [#2463] Export subsprites animation context menu on frames
 - Open in the Flash Player context menu on graphic/sound tags and frames
+- [#2451] Replacing sound stream block ranges
 
 ### Changed
 - AS1/2 - Single DoAction tag inside frame is now displayed directly as frame node
@@ -3849,6 +3850,7 @@ Major version of SWF to XML export changed to 2.
 [#2412]: https://www.free-decompiler.com/flash/issues/2412
 [#1682]: https://www.free-decompiler.com/flash/issues/1682
 [#2463]: https://www.free-decompiler.com/flash/issues/2463
+[#2451]: https://www.free-decompiler.com/flash/issues/2451
 [#2456]: https://www.free-decompiler.com/flash/issues/2456
 [#2459]: https://www.free-decompiler.com/flash/issues/2459
 [#2460]: https://www.free-decompiler.com/flash/issues/2460